Add a wrapper for the keylime tenant
The script runs the keylime tenant in a container, passing the provided
arguments.

The registrar IP can be set via the --registrar option, and the verifier
IP can be set via the --verifier option.

If not set, assume the registrar and verifier are running locally and
set their IPs with the host IP.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>